refactor: Group client and server-related stuff in proxy module
