+++ crates/extensions/tedge-p11-server/src/lib.rs
@@ -15,7 +15,7 @@ pub fn tedge_p11_service(config: CryptokiConfig) -> anyhow::Result<Arc<dyn Tedge
             Arc::new(cryptoki)
         }
         CryptokiConfig::SocketService { socket_path, uri } => {
-            let mut client = client::TedgeP11Client::with_ready_check(socket_path.into());
+            let mut client = proxy::client::TedgeP11Client::with_ready_check(socket_path.into());
             client.uri = uri;
             Arc::new(client)
         }
@@ -24,18 +24,13 @@ pub fn tedge_p11_service(config: CryptokiConfig) -> anyhow::Result<Arc<dyn Tedge
 }
 
 /// A server listening on the UNIX domain socket, wrapping the service.
-mod server;
-pub use server::TedgeP11Server;
-
-/// Serialization and framing of messages sent between the client and server.
-mod connection;
+mod proxy;
+pub use proxy::TedgeP11Client;
+pub use proxy::TedgeP11Server;
 
 /// A rustls SigningKey that connects to the server.
 mod signer;
 pub use signer::signing_key;
-
-/// A client that connects to the UNIX server, used by the signer.
-pub mod client;
 
 /// Interfaces with the PKCS#11 dynamic module using cryptoki crate.
 pub mod pkcs11;
